nbd: Switch to race-free allocation of /dev/nbdX
New enough nbd-client will use libnl to grab the next available
/dev/nbdX using netlink sockets if we omit the device name; and then
prints this name before exiting.  Let's use that instead of
hard-coding the use of /dev/nbd15.  The disconnect script is already
robust to the dynamic naming, by reading symlink contents.

Since nbd-client is already short-running (it exits as soon as the
kernel has taken over the NBD socket), the former exec in the script
was not necessary to avoid pid leaks, since the entire job container
running connect.sh is exiting quickly anyway.

Enhance the mount-rwx.sh test to create two pvcs at once, to force the
use of two distinct NBD devices when crossing nodes in the cluster.
The test change without the corresponding code change would fail when
the second nbd-client sees that /dev/nbd15 is already in use.