Don't recreate dm-multipath target upon blob pool migration
Otherwise, queued I/O is lost. We instead must modifiy the paths of the
dm-multipath target without replacing it with a new target.

It isn't possible to add or remove arbitrary paths, only to enable and
disable existing paths. We thus point dm-multipath at a "lower" dm
volume whose table we can modify to make it point to different volumes
using a dm-linear target, or none at all using a dm-error target.

Signed-off-by: Alberto Faria <afaria@redhat.com>