docs: More details on shared storage setup
Mention that the dm-thin-pool module is just as important as the nbd
module, and give a formula for setting it up.  Mention restrictions on
the backingStorePath to be passed to the StoragePath; for now, the
responsibility on wiping any earlier lvm2 PV from the storage before
deploying subprovisioner rests on the admin.

Fixes: #28
Signed-off-by: Eric Blake <eblake@redhat.com>