tests: Avoid unsafe use of xargs
Consider a git checkout in '/tmp/dir with space/subprovisioner'.
While unlikely, using run.sh from this repo would result in passing
the result of realpath to 'xargs dirname' producing the output:

/tmp
.
space/subprovisioner/tests

which is NOT what we wanted.

While at it, there's no need to fork pwd(1) when bash already
guarantees that $PWD has the same value.

Signed-off-by: Eric Blake <eblake@redhat.com>