Add LVM profile with autoextend configuration
LVM profiles are lvm.conf configuration that can be applied to specific
VGs, LVs, or command invocations. Profiles allow Subprovisioner to
specify LVM configuration settings without interfering with the
system-wide lvm.conf file.

Set the autoextend threshold to 95% of capacity so that thin pools are
automatically grown when the threshold is reached.  See lvmthin(7) for
details.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>