blob: Pass in VG, not raw block device, to StorageClass
Since we recently off-loaded the management of lvmlockd to the nodes
rather than trying to do it within containers (ie. the user must set
up the host id for each node prior to installing subprovisioner), the
next logical step is to require the user to pre-create a shared volume
group and pass that in, rather than expecting subprovisioner to take
over a bare device.

This has a couple of benefits: the user can decide whether to have
multiple VG on a single LUN ('man lvmlockd' recommends that the first
shared VG created be small and dedicated to the global lock, so that
all subsequent shared VGs can be re-homed with less effort as long as
the global lock VG is still around), or even to spread a single shared
VG across multiple lvm PV shared across the nodes.

Also, since the user is now responsible for creating the VG, we can
assume that their VG name is already unambiguous for each node of the
system, and that we no longer need to pass in '--devices /dev/XXX' to
all our lvm calls.

Fixes: https://gitlab.com/subprovisioner/subprovisioner/-/issues/36
Signed-off-by: Eric Blake <eblake@redhat.com>