tests: Avoid leaving build.tar files in /tmp
After `minikube image load -` finishes, a container image file is left
in /tmp. I noticed this when /tmp filled up on my machine.

Manage the temporary file ourselves so that cleanup is guaranteed.
Streaming behavior is lost, but it's not a real loss since minikube was
buffering data in a temporary file instead of streaming anyway.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>