docs: Show VolumeSnapshotClass creation
Snapshot creation requires a VolumeSnapshotClass. It is similar to a
StorageClass but for snapshot parameters. Subprovisioner doesn't support
any parameters but it is still necessary to define a VolumeSnapshotClass
in order to create a snapshot with the CSI plugin.

Show how to create the VolumeSnapshotClass in the Getting Started
documentation.

Fixes: #21
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>