blob: Less hardcoding of VG name 'subprovisioner'
Hide all uses of the hard-coded VG name subprovisioner behind
config.LvmVgName, to make it easier for a future patch to rely on a
per-StorageClass VG name given by the user.

Mostly mechanical, but several notes to help review:

1. 'man 8 lvm' states that /dev/vg_name/lv_name is preferred over
/dev/mapper/vg_name-lv_name, because the latter is subject to dm
munging rules, which may change between versions of dm tools

2. Although our vg_name is currently hard-coded and thus immune to dm
munging, we must do the s/-/--/g munging on the VG name anywhere that
we do need to refer to something under /dev/mapper (such as
dm-multipath, where we no longer benefit from the observation in the
first point)

3. scripts/lvm.sh now has an extra parameter of the VG name (the
parameters will change again in the next patch; the churn is
unavoidable).

Signed-off-by: Eric Blake <eblake@redhat.com>