tests: Expose two shared disks to the cluster
An upcoming patch will add a test of simultaneous StorageClass
visiting two different VGs.  The easiest way to do this is with two
NBD devices; existing tests won't use the second one, but it doesn't
hurt to set it up unconditionally.

A symlink is just fine for both backing devices; however, we also use
the vanity name /dev/my-san-lun which looks better if it is copied as
a block device (same major/minor of the nbd device) rather than a
symlink.

I did consider an alternative of partitioning a single 4G NBD shared
device, and the mapping one VG per partition; but then I'd have to
pull in fdisk or similar.  It is also possible to use qemu-nbd with
--image-opts and a raw layer on top of NBD to expose only half of the
backing device, but that also seems more complicated than just having
two backing devices.

Signed-off-by: Eric Blake <eblake@redhat.com>