docs: Show OpenShift package installation steps
OpenShift recommends running Red Hat Enterprise Linux CoreOS (RHCOS) on
nodes. RHCOS does not have lvm2-lockd or sanlock installed by default.
Show how to install it via rpm-ostree and include the steps for other
Red Hat-based distros while we're at it.

In the long term it would be appropriate for OpenShift to install the
package dependencies so that upstream Subprovisioner does not need to
address this case. That could be done in a downstream operator or maybe
by adding the required packages to the default package set. In any case,
we need to provide documentation on how to do this manually for now.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>