grpc: Add sanity checks to NodePublishVolume
The csi-sanity test [1] detected that we were not validating whether
required arguments were set, which in turn led to confusing results
(Code.Unknown instead of Code.InvalidArgument) when util.Symlink was
passed an empty string.  While a well-behaved k8s will not omit
required arguments, it never hurts to be stricter on what we accept so
that we return the correct error code.

[1] https://github.com/kubernetes-csi/csi-test/tree/master/pkg/sanity

Signed-off-by: Eric Blake <eblake@redhat.com>