deploy: make it easier to run csi-sanity
csi-sanity wants to connect to the socket for both the controller and
the node drivers.  For our node drivers, we are already exposing a
Unix socket from a known location on the host into the container, but
for the controller, we were using an empty directory which becomes
difficult to discover.  It does not hurt to name an explicit directory
for both endpoints; but with two sockets, it becomes nicer to have
consistent naming in the host.

Signed-off-by: Eric Blake <eblake@redhat.com>