build: update test regex to match latest deploy spelling
run.sh requires that we load the just-built subprovisioner:test image
from localhost, as it may contain items not present in the default
quay.io/subprovisioner tagged build mentioned in deploy.  The recent
change to pull :latest instead of an arbitrary :vN.N was not reflected
into the testsuite, which broke every test when the minikube image is
unable to access the internet to pull in snapshot support without the
proxy settings in the :test image.

Fixes: bdbcb8598258 ("build: point default images to latest tag")
Signed-off-by: Eric Blake <eblake@redhat.com>