blob: diagnose improper blob name
csi-sanity tests whether the driver can successfully ignore an attempt
to delete an unknown PVC or snapshot.  But because we weren't sanity
checking input, we triggered a go out-of-bounds panic when
dereferencing split[1] after strings.SplitN(,,3) produced only 1
string.  Kubernetes should never pass in a blob name that was not
previously learned, but this fix will let us pass a few more
csi-sanity tests.

Signed-off-by: Eric Blake <eblake@redhat.com>