Use a temporary snapshot when cloning volumes
Volume cloning is currently performed using dd(1). This is a
long-running operation during which the source volume may be modified.
The newly cloned volume can therefore contain an inconsistent copy of
the source volume (a mix of older and newer data).

Create a temporary snapshot of the source volume so that dd(1) creates a
consistent copy. Once the volume has been cloned the temporary snapshot
is deleted.

Fixes: #49
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>