tests: Fix /dev/mapper VG name prefix in migrate.sh
The migrate.sh test waits for the thin pool by looking for the
corresponding block device node in /dev/mapper. The test always fails
because the filename is outdated.

Update the test to use the name format currently in use. The LVM Volume
Group name is prefixed (with hyphens escaped) so it should be
"/dev/mapper/kubesan--vg-pvc-...".

Fixes: c3dcd4ac5602 ("blob: Pass in VG, not raw block device, to StorageClass")
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>