Do not set DM_DISABLE_UDEV environment variable
Originally the code was designed to run LVM commands from inside the
container and minimize interaction with the host's LVM and udev
infrastructure. Setting DM_DISABLE_UDEV was part of this.

It is no longer necessary to disable udev rules since the switch to the
host's LVM tools. In fact, doing so causes /dev/mapper/<vg-name>-lvmlock
to be removed by udev when a "change" uevent action is received. When
this happens (it is non-deterministic), sanlock fails to open the
lvmlock LV and the tests break.

Allow udev rules so that the /dev/mapper/<vg-name>-lvmlock symlink stays
in place across uevents.

Fixes: 61b29814d1d4 ("Use the node's lvm and sanlock")
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>