tests: Enhance t/mount-rwx.sh to read back written data
Instead of just writing all zeroes from a pod in each node (which
overwrites each other, and is not otherwise tested that what got
written persisted), change the test to call into a helper script that
intentionally writes one 64k chunk of the PVC then reads another chunk
until the neighboring node has written the expected value.

This requires tweaking the test container to include arbitrary helper
scripts (so far, test-rwx.sh is the only one to provide such a
script).  Furthermore, I could not think of any good way to keep the
container alive indefinitely for both success and failure - that's
because there's no good trigger in one pod to know when the next
node's write has completed (a timeout is liable to trigger early if
k8s does not start all of the pods in a short window of time from one
another), so it is no longer possible to './run.sh --pause-on-failure
t/mount-rwx.sh' followed by 'kubectl exec -it pods/test-pod-0 -- bash'
to get a working shell to further debug things; but hopefully the
'kubectl logs pods/test-pod-0' has enough bread-crumbs to debug what
is going on if the test fails.

Fixes: https://gitlab.com/kubesan/kubesan/-/issues/58
Signed-off-by: Eric Blake <eblake@redhat.com>