t/migrate.sh: do not ssh into node
It is easy to ssh into minikube nodes, but this is not necessarily the
case with other types of Kubernetes clusters. The test suite should
eventually run against arbitrary Kubernetes clusters. It would be
inconvenient to require ssh connection details from users (e.g. 'core'
username for OpenShift clusters and ssh key files).

Watch `dmsetup status` output in the destination test Pod to detect when
the LV appears after migration. (Using find(1) didn't work for me
because the device nodes did not appear in /dev/mapper/.)

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>