commands: preserve environment variables
The environment of the new process only has LC_ALL=C set. All other
environment variables are unset. This causes the shell to fail when
looking up commands (e.g. lvm).

I think the intention was to pass environment variables:

  cmd.Env = append(cmd.Env, "LC_ALL=C")

But cmd.Env is nil here. Use cmd.Environ() to get the current
environment. (If the intention was to use nil then it would be clearer
to write []string(nil) instead of cmd.Env, but I think that was not the
intention.)

This commit fixes lvm command invocation.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>