commands: use os.Stat() for PathExistsOnHost()
When the path does not exist, ls(1) exits with error code 2 indicating a
major error. This error is currently propagated and causes the reconcile
function to fail instead of detecting that the file does not exist. But
error code 2 could also mean other things, so ls(1) is not really suited
to testing for the existence of a path.

Thanks to `hostPID: true` it's possible to call stat directly instead of
using ls(1). This fixes the behavior of PathExistsOnHost() when the file
does not exist.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>