manager: use SetControllerReference() instead of SetOwnerReference()
The reconcile function is not being called when an owned resource is
updated.

Fix this by switching from SetOwnerReference() to
SetControllerReference(). The FatBlob controller is responsible for
LinearLV's lifecycle and the Volume controller is responsible for
FatBlob's lifecycle so they need to set controller=true. This makes
reconcile calls happen because the Owns() watcher only reconciles when
controller=true by default.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>