manager/node: make LinearLV reconcile idempotent
The following is not idempotent:

  if action_required:
      action()
      resource.Status.Done = true
      Update(resource)

The problem is that if Update() fails, then action_require is false next
time around and Update() will never be called.

Split the action from the Status update so that reconcile eventually
gets there. This makes activating LinearLV on a node work. Without this
change the resource never has Status updated.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>