csi/node: symlink device to staging target path
NodeStageVolume ensures that the block device exists at
Volume.Status.Path but does not symlink it to staging target path.
NodePublishVolume expects the block device to be at the staging target
path and attaching the volume fails as a result.

Create a symlink to set up the staging target path so that the volume
can be attached successfully.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>