Makefile: Auto-run controller-gen when *_types.go change
make is a dependency tracker, so use it to automate things that we
would otherwise have to remember to do manually!  Do this by creating
a hidden empty file whose timestamp tracks whether we need to
regenerate, so that we aren't wasting the time on builds where the
CRDs didn't change.

See also https://stackoverflow.com/questions/63770457/run-phony-targets-only-if-certain-files-change

Signed-off-by: Eric Blake <eblake@redhat.com>