volume: move fat volume code to FatBlobManager
The Volume controller in the cluster controller manager needs to support
both fat and thin volume modes. Introduce the BlobManager interface and
implement NewFatBlobManager() for fat volume (LVM linear LV) operations.
This hides the differences between fat and thin volumes from
volume_controller.go.

This is a first step at separating the Volume controller from the code
that handles the volume-specific details (like invoking LVM commands).
The code will have to evolve further to support thin-pools in the
future.

Only tackle the cluster controller manager for now. The node controller
manager will also want to use the BlobManager to perform operations on
blobs in the future. When that code is added it is likely that the
BlobManager code will be moved into its own package under
internal/manager/.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>