volume: do not attempt volume expansion in cluster controller
Volume expansion does not work in the cluster controller because it
requires the LV lock. If another node is holding the LV lock then the
volume cannot be expanded from the cluster controller manager.

Drop it for now. The long-term solution will be to perform volume
expansion in the node controller manager.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>