thinpool: add ThinPool CRD
The ThinPool CRD represents the thin-pool LV and contains fields for
coordinating activation on nodes.

Thin-pool LV creation and deletion are implemented. Volume will be
responsible for creating and deleting ThinPool CRs in the future. After
a Volume deletes a ThinPool, the ThinPool stays around until the last
remaining Volume/Snapshot belonging to the thin-pool has been deleted.
For now this lifecycle hasn't been fully implemented yet.

Activation is not implemented in this commit. It will be done in a
ThinPool node controller.

There is overlap between the existing ThinPoolLv CRD and ThinPool. I
intend to flesh out ThinPool according to the CRD design document and
then see whether the unify with ThinPoolLv.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>