api: Prefer Namespaced CRDs
Our use of CRDs is to assist the work of KubeSAN.  We don't want our
CRD names to pollute the global cluster namespace, particularly for a
CRD based on a common word like Volume.  This change means that
'kubectl get volumes' no longer works, and you now need 'kubectl
--namespace kubesan-system get volumes'; but I don't see that as a
drawback.

Related to that, our RBAC can use merely Role objects tied to the
namespace, rather than ClusterRole objects.  This gives us a bit more
security in that pods outside the namespace won't be manipulating our
resources.

Note that before the crds branch, we were using the 'kubesan'
namespace; kubebuilder defaults to giving us the 'kubesan-system'
namespace.  We may still want to tweak that in a followup patch.
Along those lines,
https://sdk.operatorframework.io/docs/building-operators/golang/operator-scope/
recommends writing the controller binary to accept the environment
variable WATCH_NAMESPACE, and setting that in
config/manager/manager.yaml to the chosen namespace, rather than
hard-coding it.  But for now, hard-coding is good enough.

Signed-off-by: Eric Blake <eblake@redhat.com>