manager: Fix finalizer Update() calls
Finalizers are part of object metadata and not part of the Status
sub-resource. Calling Status.Update() instead of a full Update() does
not write the finalizer change to the API server.

Note that this bug can be hidden by later Update() calls in the
controller, so whether or not the problem can be observed depends on the
controller implementation.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>