config: Learn namespace at runtime
Restore the use of a variable tracking the namespace learned
dynamically at process startup, rather than hard-coding it.  The
kustomization files still hardcode to "kubesan-system", but that is
because kustomize has the ability to do 'replacements' for a
structured global search-and-replace to allow for things like LVMS
wanting to deploy KubeSAN in the "openshift-storage" namespace.  This
was originally done in MR !56, but lost in the rewrite to CRDs built
by kubebuilder.

Signed-off-by: Eric Blake <eblake@redhat.com>