docs: Update to match recent CRD changes
The recent merge of !80 changed the StorageClass parameters name from
backingVolumeGroup to lvmVolumeGroup. It also added support for Linear
LVs via a new mode:Linear (with mode:Thin being the current planned
way to get old behavior, but Thin behavior is currently in a state of
flux).  The addition of CRDs has changed a bit of the design; at this
point, the design document being coded to is not yet ready for
exposure in markdown format (current develpers are meeting regularly
to refine that design, and it will eventually end up here in the
repo).

Signed-off-by: Eric Blake <eblake@redhat.com>