tests: Improve mount-rwx.sh
In commit e5243587, I claimed that there was no easy way to make the
pod stay alive for easy debugging while also getting useful status
from the test script.  But that ability to debug (in particular, the
NBD server should be present while the pod is running, and cleanly go
away after both nodes have unstaged) is going to be helpful for me on
the CRD refactoring.  Update this test to keep the pod alive by using
a second container.

Meanwhile, the script's increased backoff between iterations was
causing more delays than it was solving; trying once a second still
lets the test pass in about the same amount of time, despite me using
a hard-coded sleep 10 rather than a polling loop for the updated
jsonpath status check.

Fixes: e5243587 ("tests: Enhance t/mount-rwx.sh to read back written data")
Signed-off-by: Eric Blake <eblake@redhat.com>