crd: Clean up Volume.Status
Instead of exposing a bool, our design doc says we will add another
Status.Conditions["DataSourceCompleted"] (although at the moment, we
still don't have code in place for creating a volume from anything
other than an empty source).

There is no need to duplicate Spec.Type["Filesystem"].FsType into
Status.

Our design doc does not (yet) call out the need for Status.Path; this
string should be identical on every node, regardless of where the LV
was created, but due to escaping reasons (if the VG name includes '-'
for example) it is nice to compute it just once for later code to
refer to, which may affect whether this needs to live in Status or can
instead be a function that takes a Volume as input and produces the
name as output.

Signed-off-by: Eric Blake <eblake@redhat.com>