manager: implement thin LV creation/deletion
Thin LV creation does not work yet because the cluster controller
attempts to create/delete thin LVs while another node may have the
thin-pool activated.

Move thin LV creation/deletion into the node controller so that the
thin-pool can be activated.

This generally follows the CRD design doc v4 but tries to reuse existing
ThinPoolLv controller code as long as the differences are small.

Logging is noisy but helpful at this point. It can be reduced in the
future, if necessary.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>