crd: Rename NbdServer to NbdExport
The NbdServer CRD had no implementation, but the latest design wants
to be able to leave it up to the implementation whether we do a server
per PVC or just per Node.  Best is to match the design doc by renaming
to NbdExport, which is more generic.  Later patches will start the
implementation, which only needs a node reconciler.

Signed-off-by: Eric Blake <eblake@redhat.com>