tests: Wait for PV deletion to avoid races
When a PVC is deleted the underlying PV and CSI Volume may still live
for a while longer. Tests that run `kubectl delete pvc ...` suffer from
race conditions.

For example, when tests/run.sh uninstalls KubeSAN it deletes the
cluster-controller-manager Pod that reconciles Volume CRs. This can
cause uninstall to hang because nothing will delete the Volume once the
Pod serving as its controller has been deleted.

Introduce a ksan-delete-volume helper function that deletes the PVC and
then waits for its PV to disappear. This ensures that KubeSAN is
quiesced and it's safe to continue.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>