api: Follow OpenShift API guidelines for ThinLvStatus.State union
The OpenShift API guidelines for discriminated unions use strings to
indicate which union member is being used:
https://github.com/openshift/enhancements/blob/master/dev-guide/api-conventions.md#discriminated-unions

Refactor this now because the next commit will introduce a new Removed state.

The Creating state was never included in ThinLvState so drop it for now.

Prefix names with ThinLvStatusState instead of just ThinLvState so that
Spec can have its own State field when the next commit refactors
Spec.Activate into an enum.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>