api: Implement two-phase thin LV removal
Thin LVs are currently deleted from the thin-pool when there is a
Status.ThinLvs[] element with not corresponding Spec.ThinLvs[] element.

To be extra safe, introduce a two-phase removal process where
Spec.ThinLvs[i].State.Name must be set to "Removed". The controller will
then remove the thin LV from the thin-pool and update
Status.ThinLvs[j].State.Name to "Removed". At that point Spec.ThinLvs[i]
can be dropped.

This approach avoids accidental deletion if Spec and Status are not in
sync with the thin-pool.

Fixes: #70
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>