Release v0.5.0
Caution: this is a regression in behavior from v0.4.0 - but we aren't
promising a stable API yet.  Actions on thin volumes, such as
snapshots and cloning, are still in the process of being reimplemented
using internal CRDs rather than with the previous race-prone direct
actions.  Only Linear LV creation/deletion is intended to be supported
in this release.

Signed-off-by: Eric Blake <eblake@redhat.com>