manager: Add Thin LV support to Volume controllers
Implement Thin LV support using the ThinPoolLv CRD. This commit
implements a subset of the Volume controller reconcile steps (excluding
NBD, device-mapper, data population for cloning, etc) needed to
create/delete volumes.

Attach/detach is also implemented so that tests/t/mount-rwo.sh can run.
The test passes and has been run 100 times to check that is
deterministic.

There are two parts here: the cluster controller and the node
controller.
1. The cluster controller uses the BlobManager interface to abstract the
   main controller reconcile flow from the ThinPoolLv details.
2. The node controller implements Thin LV support directly in
   volume_node_controller.go without an abstraction like BlobManager,
   because the node reconcile flow is more specific to the type of LV
   being used (Linear or Thin).

The heart of the controller is the UpdateThinPoolLv() function, which
calculates whether the ThinPoolLv needs to be active and selects a node,
if necessary. It ensures that the ThinPoolLv is only active when
necessary.

The cluster controller and node controller code is kept separate to
avoid confusion about what is common and isn't. UpdateThinPoolLv() lives
in the new internal/manager/common/thinpoollv package. Both the cluster
controller and the node controller use this package. I chose to use this
instead of putting common code into api/v1alpha1/ because the OpenShift
API conventions forbid functions in APIs:
https://github.com/openshift/enhancements/blob/master/dev-guide/api-conventions.md#no-functions

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>