tests: Add ksan-supported-modes to skip unsupported tests
Some tests do not support both Linear and Thin LVs. Add a helper
function called ksan-supported-modes so each test case can declare which
modes it supports.

Each test case has been updated with a ksan-supported-modes call based
on the current state of the codebase. For example, any test that
requires snapshots does not support Linear or (temporarily) Thin. Thin
currently does not implement snapshots although it will in the future.
Once snapshot support is added to Thin it will be possible to update the
supported modes for that test.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>