config: Refactor kustomization files for easier overrides
LVMS wants to change the namespace for KubeSAN from 'kubesan-system'
to 'openshift-storage', as well as make it easy to pull in a fixed
version.  Make this easier to do by having only one spot in our
kustomization tree to set the release version (instead of having to
touch four different files).

While at it, note that our test suite was using a fragile sed
expression to rewrite the concatenated output of kustomization to
point to our just-built local kubesan:test image; better is to use
kustomization itself to do the rewrite in a context-sensitive manner.
Kustomize does not like absolute paths in a resource path name; but we
can (ab)use `readlink` to convert any absolute path into a relative
one (even if it goes through enough .. to reach /).

Signed-off-by: Eric Blake <eblake@redhat.com>