tests: Expose necessary paths to csi-sanity test
The csi-sanity test runs in a Pod and communicates with the CSI socket
of the KubeSAN CSI driver that is installed in the Kubernetes cluster.
The test needs access to the CSI staging and target paths. It also needs
access to /dev because the test code follows symlinks that resolve in a
device-mapper block device.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>