crd: Don't use omitempty on members we want to require
If our CEL validation requires a string to be non-empty, but we also
use ",omitempty" in the JSON description of that member, the latter
takes over and the user can still omit the element.  We introduced
this pattern by copy-and-paste of OpenShift coding guidelines for
discriminated unions, but that example has since been updated
(https://github.com/openshift/enhancements/pull/1695), so update our
code to follow suit.

Signed-off-by: Eric Blake <eblake@redhat.com>