config: Expose Pod name to kubesan executables
Upcoming patches want to spawn an NBD server using the same container
image that KubeSAN is already using for its drivers (no need to ship
two containers when one will do).  Previous patches tried to use yaml
referencing to copy-and-paste the image name into an environment
variable (see KUBESAN_IMAGE in v0.4.0; lost in refactoring back in
commit 1a90794b).  But that approach is brittle: it does NOT pick
reflect kustomization overrides to force the use of a particular image
tag, but rather reflects the hard-coded string pre-kustomize.  Since
we want our testsuite to spawn the NBD server using the kubesan:test
image, and not the kubesan:latest image, we need to get the image name
from some other means.

Kubernetes has not exposed a container's image name as one of the
items that can be exposed through DownwardAPI.  However, it does
expose both the pod name and the namespace the Pod is running in,
which is sufficient for a runtime query to grab anything else from the
pod by asking the cluster how it was created - but for this to work,
we need to actually expose the pod name to the binaries we will be
running.

Signed-off-by: Eric Blake <eblake@redhat.com>