deploy: Improve deployment for OpenShift
OpenShift prefers to stick CSI drivers in the openshift-storage
namespace, rather than our own kubesan-system.  This is an easy change
with kustomization; and earlier patches ensure that we are referencing
the preferred namespace dynamically (no hard-coded uses of
kubesan-system in the code).

Moreover, OpenShift needs a SecurityContextConstraints installed, and
this must reference the namespace as a substring of each user being
created.  Doing substring substition with a single kustomization
replacement (rather than having to a 1:1 replacement of each line of
the users list) may be possible, but it's easier to just use the
preferred namespace from the get-go.

Finally, OpenShift come with a priority class openshift-user-critical
with priority 1000000000; use this for all pods deployed by KubeSAN so
that we have priority over user Pods that might be using PVCs provided
by KubeSAN.  A later patch will then worry about also using this for
dynamic pods created at runtime.

Signed-off-by: Eric Blake <eblake@redhat.com>