workers: Add API for long-running background work
Introduce a "workers" API for running a function in a goroutine and
triggering reconcilation upon completion. This can be used to perform
long-running operations without blocking the controller.

The API is idempotent and uses the WatchPending error return to indicate
that the worker is still running:

  err := workers.Run("my-unique-id", ownerObject, args)
  if err != nil {
        return err
  }

Every time the controller calls this it returns WatchPending until the
worker is done. Then the owner object is reconciled. When Run() is
called after the worker has completed, it returns the result (nil on
success) and deletes the worker so it is started afresh next time Run()
is called.

It is necessary to think about cleanup as well. A Cancel() method allows
the controller's deletion code to cancel the worker's context.

Note that workers are local to the controller process and are not
visible at the Kubernetes CRD level. This is a disadvantage because it
prevents coordination between nodes and makes it harder to observe what
is happening (although logging can help there).

The next commit will use this API to execute blkdiscard(8) in the
background without blocking the reconcile control loop.

Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>