nbd: Use one qemu-storage-daemon process per node
NbdExport currently creates Pods to run NBD servers. There is a limit on
the number of Pods per node and new Pods cannot be scheduled when a node
is being evicted (e.g. due to memory pressure).

Switch to running a qemu-storage-daemon container in the
node-controller-manager Pod. This way each node has a long-running NBD
server where volumes can be exported/unexported as needed at runtime
without spawning new Pods.

qemu-storage-daemon is controlled through the /run/qsd/qmp.sock domain
socket. The new code assumes that commands are processed quickly by
qemu-storage-daemon so there is no need for concurrency in the
controller. This assumption is likely to hold true because
qemu-storage-daemon has its state in RAM and opening LV should be fast
(if not, then all LVM operations are likely to be slow too).

Fixes: #69
Signed-off-by: Stefan Hajnoczi <stefanha@redhat.com>